Virtual Truffle Hunting
Virtual Truffle Hunting is software created by A. Rosa-Gruszecka, D. Hilszczańska, G. Pacioni and T. Kłoszewski. The aim of this software is to enable searches for truffles (Tuber spp.), particularly the Burgundy truffle (T. aestivum Vittad.). The method has been constantly developed since 2007 by the Forest Research Institute in Poland. This method is to ascertain the truffle potential for all areas in which digital maps are not available.